pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "healthcheck-reporter"
version = "0.1.6"
description = "A health checker package to report using MQTT"
readme = "README.md"
authors = [
    {name = "Fernando González", email = "fernando.daza.dev@gmail.com"}
]
requires-python = ">=3.8"
dependencies = [
    "paho-mqtt>=1.6.1,<2.0.0",
    "psycopg2-binary>=2.9,<3.0",
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn>=0.24.0,<1.0.0",
    "requests>=2.31.0,<3.0.0"
]


[project.urls]
"Homepage" = "https://github.com/fernandodazadev/healthcheck-reporter"
"Changelog" = "https://github.com/fernandodazadev/healthcheck-reporter/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/healthcheck_reporter"]